fixed clear tasks button
