+++ code/src/App.js
@@ -5,6 +5,7 @@ import { combineReducers, configureStore } from '@reduxjs/toolkit';
 import { Info } from './components/Info'
 import { Form } from './components/Form'
 import { List } from './components/List';
+import { RemoveAll } from './components/RemoveAll';
 import { todos } from './reducers/todos';
 
 const reducer = combineReducers({ todos: todos.reducer });
@@ -16,6 +17,7 @@ export const App = () => {
             <Info />
             <Form />
             <List />
+            <RemoveAll />
         </Provider>
     );
 };
