+++ code/src/components/Item.js
@@ -5,10 +5,6 @@ import { todos } from 'reducers/todos';
 
 const ItemContainer = styled.li`
   display:flex;
-  //flex-direction:row;
-  //align-self: center;
-  //justify-content: space-around;
-  //align-items: center;
   padding: 8px;
   margin: 8px 0;
   font-size: 24px;
@@ -22,16 +18,12 @@ const ItemContainer = styled.li`
   justify-content: space-between;
 `;
 const ListRemoveButton = styled.button`
-  //margin-right:50px;
-  //align-self: flex-end;
   font-size: 3vw;
   border:none;
   color: #3f3f3f;
   background-color:#e5c5b5;
 `;
 const ListCompleteButton = styled.button`
-  //align-self: center;
-  //margin-left:50px;
   font-size: 3vw;
   color: #3f3f3f;
   border:none;
